+++ app/Http/Controllers/Admin/VehiclePartController.php
@@ -28,12 +28,15 @@ class VehiclePartController extends Controller
             'slug' => $slug,
             'is_active' => $is_active,
         ]);
-        return $this->getLatestRecords(true, 'Vehicle Part created successfully.');
+        return $this->getLatestRecords(true, 'Vehicle Part saved successfully.');
 
     }
     public function destroy($id)
     {
         $vehicle_part = VehiclePart::findOrFail($id);
+        if (!$vehicle_part) {
+            return response()->json(['success' => false, 'message' => 'Vehicle Part not found.'], 404);
+        }
         $vehicle_part->delete();
         return $this->getLatestRecords(true, 'Vehicle Part deleted successfully.');
     }
@@ -44,10 +47,11 @@ class VehiclePartController extends Controller
         return response()->json([
             'success' => true,
             'message' => 'Vehicle Part retrieved successfully.',
+            'vehiclePart' => $vehicle_part,
         ]);
     }
 
-    private function getLatestRecords($success = true , $message = 'Location created successfully.')
+    private function getLatestRecords($success = true , $message = 'vehicle saved successfully.')
     {
         $vehicle_parts = VehiclePart::latest()->get();
         $html = view('admin.vehicle-parts.data-table', compact('vehicle_parts'))->render();
